Add files via upload
Now the csv has to be modified before using the app. You must first take out the time zone information from the csv.
The dataframe now has the end and start time in datetime format so instead of getting the string and then converting it in order to use it this can be done directly.